+++ packages/medical-domain/test/repositories/statistics/insulin-statistics.service.test.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, Diabeloop
+ * Copyright (c) 2023-2024, Diabeloop
  *
  * All rights reserved.
  *
@@ -68,24 +68,28 @@ describe('getBasalBolusData', () => {
     const basals = buildBasalsData(basalsData)
     const bolus = buildBolusData(bolusData)
 
-    const basalBolusData = BasalBolusStatisticsService.getBasalBolusData(basals, bolus, 1, dateFilterOneDay)
+    const basalBolusData = BasalBolusStatisticsService.getBasalBolusData(basals, bolus, [], 1, dateFilterOneDay)
     const expectBasalBolusData = {
-      basal: 1.5,
-      bolus: 15,
-      total: 16.5
+      totalCorrectiveBolusesAndBasals: 16.5,
+      totalManualBoluses: 0,
+      totalMealBoluses: 0,
+      totalPenBoluses: 4,
+      total: 20.5
     }
     expect(basalBolusData).toEqual(expectBasalBolusData)
   })
 
-  it('should return the avg daily total basal and bolus insulin delivery when viewing more than 1 day', () => {
+  it('should return tehe avg daily total basal and bolus insulin delivery when viewing more than 1 day', () => {
     const basals = buildBasalsData(basalsData)
     const bolus = buildBolusData(bolusData)
 
-    const basalBolusData = BasalBolusStatisticsService.getBasalBolusData(basals, bolus, 2, dateFilterTwoWeeks)
+    const basalBolusData = BasalBolusStatisticsService.getBasalBolusData(basals, bolus, [], 2, dateFilterTwoWeeks)
     const expectBasalBolusData = {
-      basal: 1.25,
-      bolus: 9.5,
-      total: 10.75
+      totalCorrectiveBolusesAndBasals: 10.75,
+      totalManualBoluses: 0,
+      totalMealBoluses: 0,
+      totalPenBoluses: 2,
+      total: 12.75
     }
     expect(basalBolusData).toEqual(expectBasalBolusData)
   })